livekit-mmla/pyproject.toml
[project]
name = "livekit-mmla"
version = "0.0.1"
description = "livekit-mmla: Python client for the livekit mmla api"
authors = [
    {name = "Eduardo Davalos", email="eduardo.davalos.anaya@vanderbilt.edu"},
    {name = "Umesh Timalsina", email="umesh.timalsina@vanderbilt.edu"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">3.10"

keywords = ["multimodal", "ai", "distributed", "streaming", "real-time"]

classifiers = [
    "Programming Language :: Python :: 3"
]

dependencies = [
    "pydantic",
    "httpx",
    "pyjwt",
    "python-dotenv"
]

[project.optional-dependencies]
token-server = [
    "fastapi[all]",
    "pydantic-settings"
]

test = [
    "pytest",
    "ruff",
    "pre-commit",
    "pytest",
    "pytest-dotenv",
    "pytest-asyncio"
]


[project.urls]
homepath = "https://github.com/oele-isis-vanderbilt/ElpAgents.git"
documentation = "https://github.com/oele-isis-vanderbilt/ElpAgents.git"
repository = "https://github.com/oele-isis-vanderbilt/ElpAgents.git"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# https://setuptools.pypa.io/en/stable/userguide/datafiles.html
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]


[tool.ruff]
ignore = ["E501"]
select = ["E", "W", "F", "C", "B", "I"]
ignore-init-module-imports = true
fixable = ["I001"]  # isort fix only

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.pytest.ini_options]
env_override_existing_values = 1
env_files = ".env"
